use broadcast to distribute
